+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+# Version 0.1.2
+
+Extract JavaScript generators to NPM package called 'yambus'.
+Added Rails generator: `rails g railbus:install`.
+
 # Version 0.1.1
 
 Change prefix for `delete` routes from `destroy_` to `delete_`.